fix(security): move cross-org-link helpers out of 'use server' module (#412)
Three functions previously exported from actions/cross-org-links.ts had
no auth check:
  - flagLinksForVisibilityDrop()
  - clearLinksFlag()
  - removeLinksForConnection()

In Next.js, every export from a 'use server' module becomes an
HTTP-callable RPC endpoint. Middleware blocks unauthenticated callers,
but any logged-in user (Viewer included) could call these to:
  - flag/unflag arbitrary cross-org links across the entire instance
  - mass-delete cross-org links between any two orgs by guessing UUIDs
  - inject attacker-controlled flagReason text into the DB
  - produce audit rows with no userId

Fix: relocate the helpers to lib/cross-org-link-helpers.ts. Files in
lib/ are not 'use server' modules, so the helpers cannot be reached as
RPC endpoints by construction. Each call site (editCapability,
editPersona, removeConnection) is itself an authenticated mutation
that has already validated the actor's session.

removeLinksForConnection now requires actorUserId and actorOrgId
parameters so the audit row identifies the user who performed the
removal — closing the "audit row with no userId" gap noted alongside
this issue.

Test added (tests/integration/cross-org-link-helpers.test.ts):
  - helpers are exported from lib/cross-org-link-helpers
  - actions/cross-org-links no longer exports the helpers (locks in
    the architecture; if anyone re-exports them from actions/, this
    fails)

All 181 integration tests pass; typecheck clean.

Capability: cms/multi-org
Closes #412
